hotelweb/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
email_validator
python-dotenv
